src/auto_test/tools/prepare_api/requirement.txt
Click==7.0
cymysql==0.9.13
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
futures==3.2.0
httplib2==0.13.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyYAML==5.1.1
SQLAlchemy==1.3.5
Werkzeug==0.15.4
